[SYN1] Moved OpenGL API class to a better suiting directory. Also added implementations for creating buffers, shaders and vertex arrays.
